LPD-94800 Sort input variables and consolidate agent tests
Address review feedback on the Auto Categorize and Generate Tags agents. The inputVariables in the agent definitions and the input-variables in both workflow definitions are now sorted alphabetically to follow the existing pattern. The AutoCategorizeAgentTest and GenerateTagsAgentTest classes are folded into AgentInstanceResourceTest as private helpers invoked from testPostAgentInstance, mirroring _testPostAgentInstanceWithTypeMakeShorter, so every agent-instance scenario lives under one umbrella test.
