+++ modules/dxp/apps/ai-hub/ai-hub-site-initializer/src/main/resources/site-initializer/object-entries/agent-definition-object-entries.json
@@ -4,7 +4,7 @@
 			"active": true,
 			"description": "This agent analyzes content and proposes categories drawn only from the vocabularies in scope. It never creates new categories.",
 			"externalReferenceCode": "L_AUTO_CATEGORIZE",
-			"inputVariables": "content,candidateCategories,count",
+			"inputVariables": "candidateCategories,content,count",
 			"outputVariable": "suggestedCategories",
 			"r_accountToAIHubAgentDefinitions_accountEntryERC": "L_AI_HUB",
 			"system": true,
@@ -43,7 +43,7 @@
 			"active": true,
 			"description": "This agent analyzes content and proposes tags, prioritizing reuse of existing tags and proposing new tags only when needed.",
 			"externalReferenceCode": "L_GENERATE_TAGS",
-			"inputVariables": "content,existingTags,count",
+			"inputVariables": "content,count,existingTags",
 			"outputVariable": "suggestedTags",
 			"r_accountToAIHubAgentDefinitions_accountEntryERC": "L_AI_HUB",
 			"system": true,
