+++ modules/dxp/apps/ai-hub/ai-hub-site-initializer/src/main/resources/site-initializer/workflow-definitions/auto-categorize-workflow-definition/workflow-definition.xml
@@ -28,11 +28,11 @@
 			<![CDATA[
 [
 	{
-		"name": "content",
+		"name": "candidateCategories",
 		"type": "string"
 	},
 	{
-		"name": "candidateCategories",
+		"name": "content",
 		"type": "string"
 	},
 	{
