+++ modules/dxp/apps/ai-hub/ai-hub-site-initializer/src/main/resources/site-initializer/workflow-definitions/generate-tags-workflow-definition/workflow-definition.xml
@@ -32,11 +32,11 @@
 		"type": "string"
 	},
 	{
-		"name": "existingTags",
+		"name": "count",
 		"type": "string"
 	},
 	{
-		"name": "count",
+		"name": "existingTags",
 		"type": "string"
 	}
 ]
